requirements.txt
asgiref==3.4.1
click==8.0.1
Django==3.2.5
django-crispy-forms==1.12.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
gunicorn==20.1.0
h11==0.12.0
Pillow==8.3.1
pytz==2021.1
sqlparse==0.4.1
uvicorn==0.14.0
